Extract unit test for IngestJobMessageHandler
